fix: restore mini-program interactive selector defaults
